arbitrage_opp/requirements.txt
python-crontab===2.5.1
aiohttp==3.7.4.post0
astroid==2.3.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
autopep8==1.5.7
backports.ssl-match-hostname==3.7.0.1
chardet==4.0.0
colored==1.4.2
gevent==21.1.2
greenlet==1.0.0
idna==3.1
iniconfig==1.1.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pylint==2.4.4
pyparsing==2.4.7
pytest==6.2.4
pytest-dotenv==0.5.2
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
websocket==0.2.1
websocket-client==0.58.0
websockets==9.1
wrapt==1.11.2
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
psycopg2==2.9.1
